APSs/aps3/requirements.txt
boto3==1.35.34
botocore==1.35.34
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10
jmespath==1.0.1
joblib==1.4.2
lightgbm==4.5.0
numpy==2.1.2
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
s3transfer==0.10.2
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
threadpoolctl==3.5.0
tzdata==2024.2
urllib3==2.2.3
